refactor: front-api -> client-api
