--- src/lib/client-api/api.interface.ts
@@ -1,6 +1,0 @@
-export interface CartInfo {
-  id: string;
-  productIdOnStore: string;
-  productPrice: number;
-  quantity: number;
-}
