--- src/lib/front-api/index.ts
@@ -1,3 +1,0 @@
-export type { FrontAPI } from "./front-api.interface";
-
-export { Cafe24API } from "./cafe24";
